requirements.txt
torch
torchmetrics
numpy
pandas
torchvision
scikit-learn
jupyterlab
matplotlib
pillow
lightning
matplotlib
opencv-python
seaborn
tensorboard
balanced-loss
adversarial-robustness-toolbox
umap-learn
grad-cam
shap
lime
scikit-image